Move AnyNetwork types into their own crates
