+++ crates/network/src/lib.rs
@@ -24,8 +24,8 @@ pub use ethereum::{Ethereum, EthereumWallet};
 mod any;
 pub use any::{
     AnyHeader, AnyNetwork, AnyReceiptEnvelope, AnyRpcBlock, AnyRpcHeader, AnyRpcTransaction,
-    AnyTransactionReceipt, AnyTxEnvelope, AnyTxType, AnyTypedTransaction, UnknownTxEnvelope,
-    UnknownTypedTransaction,
+    AnyTransactionReceipt, AnyTransactionRequest, AnyTxEnvelope, AnyTxType, AnyTypedTransaction,
+    UnknownTxEnvelope, UnknownTypedTransaction,
 };
 
 pub use alloy_eips::eip2718;
