+++ crates/provider/src/provider/trait.rs
@@ -1732,7 +1732,8 @@ mod tests {
 
     #[tokio::test]
     async fn any_network_wallet_filler() {
-        use alloy_serde::WithOtherFields;
+        use alloy_rpc_types_any::AnyTransactionRequest;
+
         let anvil = Anvil::new().spawn();
         let signer: PrivateKeySigner =
             "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80".parse().unwrap();
@@ -1748,7 +1749,7 @@ mod tests {
             .with_to(address!("c02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"))
             .value(U256::from(325235));
 
-        let tx = WithOtherFields::new(tx);
+        let tx = AnyTransactionRequest::new(tx);
 
         let builder = provider.send_transaction(tx).await.unwrap().get_receipt().await.unwrap();
