+++ crates/rpc-types-any/Cargo.toml
@@ -20,11 +20,14 @@ workspace = true
 
 [dependencies]
 alloy-consensus-any = { workspace = true, features = ["serde"] }
+alloy-eips.workspace = true
+alloy-network-primitives.workspace = true
 alloy-rpc-types-eth.workspace = true
 alloy-serde.workspace = true
 
-[dev-dependencies]
-alloy-primitives.workspace = true
-
 serde.workspace = true
 serde_json.workspace = true
+
+[dev-dependencies]
+alloy-consensus = { workspace = true, features = ["serde"] }
+alloy-primitives.workspace = true
